snap/hooks/install, snap/snapcraft.yaml: configure default config file with sed
During the install hook when the snap is being installed we will generate the config file from the default one that is provided in the kong repo. We make 3 replacements:
- the prefix is set to $SNAP_DATA
- the nginx user is set to root
- the database is set to cassandra
Note that a user can always modify these config values if they want or use a different kong config file entirely, this just provides a usable one directly with the snap.