many: do postgres setup in post-refresh hook too
We also need to do the postgres setup in the post-refresh hook (in addition to
the install hook) so that users who are upgraded from the current version which
only uses Cassandra can seamlessly switch to PostgreSQL if they so choose.
Otherwise, they would have to copy config files and manually process them, etc.
which is a poor user experience.

Add a note about this to the README.md, including how to migrate from Cassandra
to PostgreSQL if a user desires.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>